feat: 다운로드 토큰 발급 API 컨버터 구현

+++ src/main/java/umc/cockple/demo/domain/chat/dto/ChatDownloadTokenDTO.java
@@ -9,7 +9,7 @@ public class ChatDownloadTokenDTO {
     public record Response(
             String downloadToken,
             String downloadUrl,
-            Long expiresIn,
+            int expiresIn,
             LocalDateTime expiresAt
     ) {}
 }
